backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
pydantic==2.11.3
PyYAML==6.0.2
python-multipart==0.0.20
pytest==8.3.5
pytest-bdd==8.1.0
httpx==0.28.1
aiofiles==24.1.0
rapidfuzz==3.13.0
